requirements.txt
absl-py==2.2.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.2
celery==5.3.4
certifi==2024.12.14
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
crashtest==0.4.1
crispy-bootstrap5==2023.10
cryptography==44.0.1
cssselect2==0.8.0
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.1.8
django-axes==6.0.3
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-crispy-forms==2.1
django-cronjobs==0.2.3
django-debug-toolbar==5.1.0
django-dirtyfields==1.9.7
django-extensions==4.1
django-filter==25.1
django-fsm==2.8.1
django-guardian==2.4.0
django-js-asset==3.1.2
django-plotly-dash==2.4.6
django-ratelimit2==0.4.2
django-redis==5.4.0
django-reversion==5.0.8
django-rq==2.8.1
django-scopes==2.0.0
django-simple-history==3.3.0
django-storages==1.14.6
django-tenants==3.7.0
django-viewflow==2.2.10
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
dpd_components==0.1.0
drf-nested-routers==0.94.1
drf-yasg==1.21.7
duckduckgo_search==8.0.0
dulwich==0.22.7
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2024.12.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
installer==0.7.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.2
keyring==25.6.0
kiwisolver==1.4.8
kombu==5.4.2
libclang==18.1.1
lxml==5.3.1
Markdown==3.8
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
msgpack==1.1.0
namex==0.0.8
narwhals==1.35.0
nest-asyncio==1.6.0
node==1.2.2
numpy==2.1.3
odict==1.9.0
openai==1.77.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pipdeptree==2.26.0
pkginfo==1.12.0
platformdirs==4.3.6
plotly==6.0.1
plumber==1.7
poetry==2.0.1
poetry-core==2.0.1
primp==0.14.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==5.29.4
psycopg==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyClamd==0.4.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
pyphen==0.17.2
pyproject_hooks==1.2.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.0
python-slugify==8.0.1
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.1
RapidFuzz==3.11.0
redis==5.0.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rpds-py==0.24.0
rq==2.3.2
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.26.1
serializers==0.2.4
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smolagents==1.12.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
svglib==1.5.1
termcolor==3.0.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.1.15.22
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.1
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.0.6
wheel==0.45.1
wrapt==1.17.2
xhtml2pdf==0.2.17
zipp==3.21.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.0
zope.proxy==6.1
zopfli==0.2.3.post1
